mcp sample cleaned up
