+++ samples/dotnet/mcp/Agents/Agent.Core/Protos/YaapClientDetail_gRPC.cs
@@ -1,14 +1,14 @@
 ﻿namespace Grpc.YaapServer;
 public partial class YaapClientDetail
 {
-    public static implicit operator Yaap.Models.YaapClientDetail(YaapClientDetail grpcYaapClientDetail) =>
+    public static implicit operator Yaap.Core.Models.YaapClientDetail(YaapClientDetail grpcYaapClientDetail) =>
         new(
             grpcYaapClientDetail.Name,
             grpcYaapClientDetail.Description,
             string.IsNullOrWhiteSpace(grpcYaapClientDetail.CallbackUrl) ? null : new(grpcYaapClientDetail.CallbackUrl)
         );
 
-    public static implicit operator YaapClientDetail(Yaap.Models.YaapClientDetail yaapClientDetail) =>
+    public static implicit operator YaapClientDetail(Yaap.Core.Models.YaapClientDetail yaapClientDetail) =>
         new()
         {
             Name = yaapClientDetail.Name,
