+++ samples/dotnet/mcp/Agents/Agent.Core/Protos/expert.proto
