+++ samples/dotnet/mcp/Agents/Agent.Core/Protos/models.proto
