+++ samples/dotnet/mcp/Agents/Agent.Core/Protos/yaap_server.proto
