+++ samples/dotnet/mcp/Agents/Orchestrator/GrpcServices/OrchestratorService.cs
@@ -1,4 +1,6 @@
-namespace Orchestrator_gRPC.GrpcServices;
+namespace Orchestrator.GrpcServices;
+
+using Agent.Core;
 
 using Google.Protobuf.WellKnownTypes;
 
@@ -6,15 +8,9 @@ using Grpc.Core;
 using Grpc.Models;
 using Grpc.Orchestrator;
 
-using gRPCAgent.Core;
-
 using Microsoft.SemanticKernel;
 
-using Orchestrator_gRPC;
-
-using Yaap.Server;
-
-internal class OrchestratorService(ILogger<OrchestratorService> _log, Kernel _kernel, PromptExecutionSettings _promptSettings) : Grpc.Orchestrator.Orchestrator.OrchestratorBase
+internal class OrchestratorService(ILogger<OrchestratorService> _log, Kernel _kernel, PromptExecutionSettings _promptSettings) : Orchestrator.OrchestratorBase
 {
     public override Task<Empty> Message(MessageRequest request, ServerCallContext context)
     {
