+++ samples/dotnet/mcp/Agents/TeamsExpert/TeamApi_Extras.cs
