+++ samples/dotnet/mcp/Agents/TeamsExpert/appsettings.Development.json
