--- samples/dotnet/mcp/Agents/gRPC/gRPCAgent.Core/gRPCAgent.Core.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<ItemGroup>
-	  <None Remove="Protos\models.proto" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Azure.Identity" />
-		<PackageReference Include="Microsoft.SemanticKernel" />
-		<PackageReference Include="Grpc.AspNetCore" />
-		<PackageReference Include="System.Linq.Async" />
-		<PackageReference Include="ModelContextProtocol" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\..\..\..\..\..\src\Client\Yaap.Client.csproj" />
-		<ProjectReference Include="..\..\..\..\mcp\Models\Models.csproj" />
-		<ProjectReference Include="..\..\..\ApiClient\ApiClient.csproj" />
-		<ProjectReference Include="..\..\..\Assistants\Assistants.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <Protobuf ProtoRoot="Protos" Include="Protos\expert.proto" GrpcServices="Both" Access="Public" ProtoCompile="True" />
-	  <Protobuf Include="Protos\models.proto" GrpcServices="None" />
-	  <Protobuf ProtoRoot="Protos" Include="Protos\orchestrator.proto" Access="Public" GrpcServices="Both" />
-	  <Protobuf Include="Protos\yaap_server.proto" />
-	</ItemGroup>
-
-</Project>
