+++ samples/dotnet/mcp/ConsoleApp/ConsoleApp.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Agents\gRPC\gRPCAgent.Core\gRPCAgent.Core.csproj" />
+    <ProjectReference Include="..\Agents\Agent.Core\Agent.Core.csproj" />
     <ProjectReference Include="..\Common\Common.csproj" />
     <ProjectReference Include="..\Models\Models.csproj" />
   </ItemGroup>
