+++ samples/dotnet/mcp/ConsoleApp/Log.cs
