+++ samples/dotnet/mcp/ConsoleApp/Program.cs
