+++ samples/dotnet/mcp/ConsoleApp/Properties/launchSettings.json
