+++ samples/dotnet/mcp/ConsoleApp/Worker.cs
