+++ samples/dotnet/mcp/ConsoleApp/appsettings.Development.json
