+++ samples/dotnet/mcp/ConsoleApp/appsettings.json
