--- samples/dotnet/mcp/TBAStatReader_WS/appsettings.json
@@ -1,10 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "None"
-    }
-  },
-
-  "McpServerEndpoint": "ws://localhost:5172/ws/orchestrator",
-  "ExpertWaitTimeSeconds": 60
-}
