--- samples/dotnet/mcp/TBAStatReader_gRPC/Properties/launchSettings.json
@@ -1,11 +1,0 @@
-{
-  "profiles": {
-    "WSL": {
-      "commandName": "WSL2",
-      "distributionName": ""
-    },
-    "ConsoleApp": {
-      "commandName": "Project"
-    }
-  }
-}
