--- samples/dotnet/mcp/TBAStatReader_gRPC/appsettings.Development.json
@@ -1,14 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      //"ConsoleApp": "Trace",
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning",
-      "Microsoft.Azure": "Warning",
-      "Microsoft.SemanticKernel": "Warning",
-      "System.Net.Http": "Warning"
-    }
-  },
-
-  "ExpertWaitTimeSeconds": 300
-}
