+++ samples/dotnet/mcp/Yaap-MCP.slnLaunch
@@ -8,19 +8,19 @@
         "DebugTarget": "AgentMcpServer"
       },
       {
-        "Path": "Agents\\gRPC\\Orchestrator_gRPC\\Orchestrator_gRPC.csproj",
+        "Path": "Agents\\Orchestrator\\Orchestrator.csproj",
         "Action": "Start",
         "DebugTarget": "http"
       },
       {
-        "Path": "TBAStatReader_gRPC\\ConsoleApp.csproj",
+        "Path": "Agents\\TeamsExpert\\TeamsExpert.csproj",
         "Action": "Start",
-        "DebugTarget": "ConsoleApp"
+        "DebugTarget": "TeamsExpert"
       },
       {
-        "Path": "Agents\\gRPC\\Teams_gRPC\\Teams_gRPC.csproj",
+        "Path": "ConsoleApp\\ConsoleApp.csproj",
         "Action": "Start",
-        "DebugTarget": "Teams_gRPC"
+        "DebugTarget": "ConsoleApp"
       }
     ]
   }
